Subnautica: move mod exports to own module
